Adds RecentLinkItem component. Fixes typo.
