+++ src/components/RecentLinks.jsx
@@ -2,17 +2,11 @@ import React, { Fragment } from 'react';
 
 import Col from 'react-bootstrap/Col';
 import Row from 'react-bootstrap/Row';
+import RecentLinkItem from './RecentLinkItem'
 
-import { graphql } from 'gatsby';
+export default ({ links, showAmount }) => {
 
-export default ({ links, showAmmount }) => {
-
-    const recentLinks = links.slice((-1 * showAmmount))
-
-    const shorten = (str, maxLen, separator = ' ') => {
-        if (str.length <= maxLen) return str;
-        return str.substr(0, str.lastIndexOf(separator, maxLen)) + '...';
-    }
+    const recentLinks = links.slice((-1 * showAmount))
 
     return (
         <div>
@@ -20,14 +14,10 @@ export default ({ links, showAmmount }) => {
                 <p className="h2"> Recent Links</p>
             </div>
             {
-                recentLinks.map(link => (
-                    <Fragment>
-                        <a href={link.url} target="_blank" class="py-3 pl-8 pr-2 border-bottom d-block text-decoration-none">
-                            <p className="h4"> {link.title} </p>
-                            <div> {shorten(link.description, 90)} </div>
-                        </a>
-                    </Fragment>
-                ))
+                recentLinks.map((link, index) => {
+                    return <RecentLinkItem key={index} link={link} />
+                }
+                )
             }
         </div>
